% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotting.R
\name{altmanbland}
\alias{altmanbland}
\title{altman-bland plot}
\usage{
altmanbland(x, y, main = "", pch = ".", log = "")
}
\arguments{
\item{x}{- input intensities}

\item{y}{-}

\item{main}{plotting parameters}

\item{pch}{-}

\item{log}{-}
}
\description{
plots abs(x-y) against (y+x)/2
}
\examples{
data(SDat)
altmanbland(SDat$Intensity[,1],SDat$Intensity[,2],log="xy")
}

