% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/analyzeDuplicatedPeptides.R
\name{analyzeDuplicated}
\alias{analyzeDuplicated}
\title{analyse duplicated peptides/proteins}
\usage{
analyzeDuplicated(data, dups, maxpep = 3, countmax = 1000,
  method = "spearman")
}
\arguments{
\item{data}{- msExperiment}

\item{dups}{- list as returned by function \code{\link{getListOfDuplicated}}}

\item{maxpep}{- limit the number of duplicated peptides}

\item{countmax}{- limit the total number of comparisons}

\item{method}{- type of correlation to compute}
}
\description{
analyse duplicated peptides/proteins
}
\examples{
data(SDat)
rownames = SDat$pepinfo$PeptideSequence
dups = getListOfDuplicated(rownames)

res = analyzeDuplicated(SDat , dups[1:25])
dim(res)
res[1,]
hist(res$cor)
plot(res$medianRTDiff,res$cor)

rownames = SDat$pepinfo$PeptideSequence
nondups = getUniqueEntries(rownames)
length(nondups)
if(length(nondups > 0)){
 res = analyzeDuplicated(SDat , list(nondups[1:25]))
}
}
\seealso{
\code{\link{getListOfDuplicated}} \code{\link{analyzeDuplicatedPeptides}}
}

