% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/analyzeDuplicatedPeptides.R
\name{analyzeDuplicatedPeptides}
\alias{analyzeDuplicatedPeptides}
\title{analyse duplicated peptides - same peptide different charge}
\usage{
analyzeDuplicatedPeptides(data, countmax = 1000)
}
\arguments{
\item{data}{msExperiment}

\item{countmax}{maximum number of duplicate comparisons}
}
\description{
analyse duplicated peptides - same peptide different charge
}
\examples{
data(SDat)
res = analyzeDuplicatedPeptides(SDat,countmax= 20)
res[1,]
hist(res$cor)
plot(res$medianRTDiff,res$cor)
#same for proteins
rownames = as.character(SDat$pepinfo$ProteinName)
}
\seealso{
\code{\link{getListOfDuplicated}} \code{\link{analyzeDuplicatedProteins}}
}

