% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/analyzeDuplicatedPeptides.R
\name{analyzeDuplicatedProteinsTOP}
\alias{analyzeDuplicatedProteinsTOP}
\title{analyse duplicated protein peptides - same protein id different peptide take top peptides}
\usage{
analyzeDuplicatedProteinsTOP(data, maxpep = 3, countmax = 1000)
}
\arguments{
\item{data}{an object of class msexperiment}

\item{maxpep}{maximum number of peptides to compare}

\item{countmax}{maximum number of duplicate comparisons}
}
\description{
analyse duplicated protein peptides - same protein id different peptide take top peptides
}
\examples{
data(SDat)
res = analyzeDuplicatedProteinsTOP(SDat)
res[1,]
hist(res$cor)
plot(res$medianRTDiff,res$cor)
#same for proteins
rownames = as.character(SDat$pepinfo$ProteinName)
}
\seealso{
\code{\link{getListOfDuplicated}} \code{\link{analyzeDuplicatedPeptides}}

\code{\link{analyzeDuplicated}}
}

