% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyzeDuplicatedPeptides.R
\name{compPeptideCorrelations}
\alias{compPeptideCorrelations}
\title{compute various correlation}
\usage{
compPeptideCorrelations(specLib, countmax = 500)
}
\arguments{
\item{specLib}{msexperiment}

\item{countmax}{max number of comparisons}
}
\description{
\itemize{
\item random - Features/Peakgroupscan be randomly drawn from the dataset.
\item unique - Features/Peakgroups can be randomly drawn among those peptides which uniquely represent proteins.
\item protein - correlation among peakgroups derived from the same protein can be computed.
\item proteinTOP - The correlation among peaks derived from the same protein and having good intensities.
\item peptide - correlation of peaks representing the same peptide but with different charge.
}
}
\examples{
data(SDat)
specLib = orderByRT(SDat)
pairsDifference(specLib$Intensity,specLib$RT)

res = compPeptideCorrelations(specLib,countmax=100)

}

