% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swathDataUtils.R
\name{convert2msExperiment}
\alias{convert2msExperiment}
\title{converts to msExperiment (exported more for debugging purpose)}
\usage{
convert2msExperiment(data)
}
\arguments{
\item{data}{output of convertLF2Wideformat}
}
\description{
converts to msExperiment (exported more for debugging purpose)
}
\examples{
data(feature_alignment_requant) 
data = convertLF2Wideformat(feature_alignment_requant)
lapply(data,dim)
SDat=convert2msExperiment(data)
stopifnot(rownames(SDat$pepinfo)==rownames(SDat$Intensity))
stopifnot(dim(SDat$Intensity) == dim(SDat$mz), dim(SDat$mz) == dim(SDat$score))
}
\seealso{
\code{\link{read2msExperiment}} and \code{\link{convertLF2Wideformat}} for contex
}

