% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/convert2msstats.r
\name{convert2MSstats}
\alias{convert2MSstats}
\title{convert to MSstats}
\usage{
convert2MSstats(msexp)
}
\arguments{
\item{msexp}{to convert}
}
\description{
convert msexp to msstats compatible output
}
\examples{
data(feature_alignment_requant)
SDat = loadTransitonsMSExperiment(feature_alignment_requant)
msstatsformat = convert2MSstats(SDat)
}

