% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/swathDataUtils.R
\name{convertLF2Wideformat}
\alias{convertLF2Wideformat}
\title{convert LF 2 wide format}
\usage{
convertLF2Wideformat(aligtable)
}
\arguments{
\item{aligtable}{output of RT alignment tool (long format)}
}
\description{
convert LF 2 wide format
}
\examples{
data(feature_alignment_requant)
feature_alignment_requant = prepareDF(feature_alignment_requant)
SDat =  convertLF2Wideformat(  feature_alignment_requant  )
}
\seealso{
\code{\link{convert2msExperiment}} for contex
}

