% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/correctIntRT.R
\name{correctIntRTv1.default}
\alias{correctIntRTv1.default}
\title{correct intensity over RT for entire msexperiment}
\usage{
\method{correctIntRTv1}{default}(obj, data, rto, plot = TRUE, k = 501,
  FUN = function(x, k = k) {     runmed(x, k = k, endrule = "constant") },
  ...)
}
\arguments{
\item{obj}{reference data - ordered by retention time}

\item{data}{to correct}

\item{rto}{retention time (ordered)}

\item{plot}{if TRUE create debug plot}

\item{k}{smoothing with}

\item{FUN}{running function - for equal median use runmed, for equal max use i.e. runFUN(x,k=300,max)}
}
\value{
corrected data
}
\description{
correct intensity over RT for entire msexperiment
}
\examples{
data(SDat)
experiment = removeDecoys(SDat)
experiment = orderByRT(experiment)
cor = correctIntRTv1(experiment$Intensity[,1], experiment$Intensity[,2],experiment$RT,k=51)
cor = correctIntRTv1(experiment$Intensity[,1], experiment$Intensity[,2],experiment$RT,k=51,FUN=function(x,k=k){print(k);runFun(x,k=k,mean)})
}
\author{
Witold Wolski \email{wolski@gmail.com}
}

