% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/correctIntRT.R
\name{correctIntRTv1.matrix}
\alias{correctIntRTv1.matrix}
\title{correct intensity over RT for entire msexperiment}
\usage{
\method{correctIntRTv1}{matrix}(obj, rt, k = 501, ...)
}
\arguments{
\item{obj}{a matrix}

\item{rt}{retention time}

\item{k}{smoothing with (data points)}

\item{...}{(unused)}
}
\description{
correct intensity over RT for entire msexperiment
}

