% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/correctIntRT.R
\name{correctIntRTv1.msexperiment}
\alias{correctIntRTv1.msexperiment}
\title{correct intensity over RT for entire msexperiment}
\usage{
\method{correctIntRTv1}{msexperiment}(obj, k = 501, ...)
}
\arguments{
\item{obj}{- object of class msexperiment}

\item{k}{- smoothing with}

\item{...}{(unused)}
}
\value{
msexperiment object with RT normalized intensities
}
\description{
correct intensity over RT for entire msexperiment
}
\note{
reorders all entries in experiment according to RT, finds sample with fewest NA's if there are many picks that one
with largest median as reference.
}
\examples{
data(SDat)
SDatal = correctIntRTv1(SDat,plot=F)
}
\author{
Witold Wolski \email{wolski@gmail.com}
}

