% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/correctIntRT.R
\name{correctIntRTv1}
\alias{correctIntRTv1}
\title{correct intensity over RT for entire msexperiment}
\usage{
correctIntRTv1(obj, ...)
}
\arguments{
\item{obj}{object}

\item{...}{parameters for specialized functions}
}
\description{
correct intensity over RT for entire msexperiment
}

