% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/correctIntRTv2.R
\name{correctIntRTv2.default}
\alias{correctIntRTv2}
\alias{correctIntRTv2.default}
\title{correct Intensity RT using func, i.e. runrobscale (default)}
\usage{
\method{correctIntRTv2}{default}(obj, rto, k = 501, scale = FALSE,
  plot = TRUE, func = runrobscale, ...)
}
\arguments{
\item{obj}{to correct}

\item{rto}{retention time}

\item{k}{smoothing with (see runmed)}

\item{scale}{should scaling be applied}

\item{plot}{show diagnostic plot}

\item{func}{function to apply for scaling see runrobscale}
}
\value{
corrected data
}
\description{
correct Intensity RT using func, i.e. runrobscale (default)
}
\examples{
res = c(rnorm(1000,0,1),rnorm(2000,0,3))
res[sample(1:length(res),100)] = NA
rto = as.numeric(1:length(res))
res2 = correctIntRTv2(res ,rto )
}
\author{
Witold Wolski
}
\seealso{
runrobscale runTICscale
}

