% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correctIntRTv2.R
\name{correctIntRTv2.msexperiment}
\alias{correctIntRTv2.msexperiment}
\title{correct intensity over RT for entire msexperiment}
\usage{
\method{correctIntRTv2}{msexperiment}(obj, k = 501, scale = FALSE,
  func = runrobscale, plot = FALSE, ...)
}
\arguments{
\item{obj}{object of class msexperiment}

\item{k}{smoothing with}

\item{scale}{should scaling also be applied}

\item{func}{function to use fore scaling: \code{\link{runrobscale}} or \code{\link{runTICscale}}}

\item{plot}{plot}

\item{...}{unused}
}
\value{
msexperiment object with RT normalized intensities
}
\description{
correct intensity over RT for entire msexperiment
}
\note{
reorders all entries in experiment according to RT, finds sample with fewest NA's if there are many picks that one
with largest median as reference.
}
\examples{
data(SDat)
res = correctIntRTv2(SDat)

}
\author{
Witold Wolski \email{wolski@gmail.com}
}

