% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/swathDataUtils.R
\name{dimnames.msexperiment}
\alias{dimnames.msexperiment}
\title{show colnames (it does'nt let you set the columns)}
\usage{
\method{dimnames}{msexperiment}(x)
}
\arguments{
\item{x}{msexperiment}
}
\description{
show colnames (it does'nt let you set the columns)
}
\examples{
data(feature_alignment_requant)
SDat = read2msExperiment(feature_alignment_requant)
dimnames(SDat)
}

