% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/hello.R
\docType{data}
\name{feature_alignment_requant}
\alias{feature_alignment_requant}
\title{feature_alignment_requant - output of feature aligner requant}
\format{see example below}
\usage{
feature_alignment_requant
}
\description{
this is an example output when running feature aligner requant on swath data
}
\author{
Lorenz Blum
}
\seealso{
\code{\link{convertLF2Wideformat}} and \code{\link{read2msExperiment.data.frame}}
data(feature_alignment_requant)
colnames(feature_alignment_requant)
head(feature_alignment_requant[,1:8])
}

