% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/filterbyScore.R
\name{filterByScore}
\alias{filterByScore}
\title{filter rows by qscore}
\usage{
filterByScore(msexp, scorethresh = 0.001, n = 2)
}
\arguments{
\item{scorethresh}{- score threshold}

\item{n}{- number of samples which need to have a score for this feature below the threshold}

\item{object}{of class msexperiment}
}
\description{
require that at least n features have a mscore less than maxscore
}
\examples{
NULL
}

