% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyzeDuplicatedPeptides.R
\name{getListOfDuplicated}
\alias{getListOfDuplicated}
\title{find duplicated entries in vector}
\usage{
getListOfDuplicated(dataw)
}
\arguments{
\item{dataw}{data}
}
\value{
list with indices of duplicated entries
}
\description{
find duplicated entries in vector
}
\examples{
data(SDat)
tmp = c("a","b","a","d","d","e","f")
res = getListOfDuplicated(tmp)
stopifnot(tmp[res[[1]]] == c("a","a"))
stopifnot(tmp[res[[2]]] == c("d","d"))

rownames = SDat$pepinfo$PeptideSequence
res = getListOfDuplicated(rownames)
rownames[res[[1]]]
rownames = as.character(SDat$pepinfo$ProteinName)
res = getListOfDuplicated(rownames)
stopifnot((unlist(lapply(res,length)) >= 2) == TRUE)
rownames[res[[1]]]
}
\seealso{
\code{\link{analyzeDuplicated}} for contex
}

