% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/analyzeDuplicatedPeptides.R
\name{getListOfDuplicatedOrderedByIntensity}
\alias{getListOfDuplicatedOrderedByIntensity}
\title{get a list of duplicates (their indices) ordered by intensity}
\usage{
getListOfDuplicatedOrderedByIntensity(rownames, intensities)
}
\description{
get a list of duplicates (their indices) ordered by intensity
}
\examples{
nam = c('a','a','b','e','c','d','f','e','e')
intens = c(1,2 , 1 , 1 , 2 , 3 , 4 , 2 , 0)
res= getListOfDuplicatedOrderedByIntensity(nam,intens)
nam[res[[1]]]
intens[res[[1]]]
nam[res[[2]]]
intens[res[[2]]]
data(SDat)
rownames = SDat$pepinfo$ProteinName
intens = apply(SDat$Intensity,1,mean)
res = getListOfDuplicatedOrderedByIntensity(rownames, apply(SDat$Intensity,1,mean))
res
intens[res[[81]]]
rownames[res[[81]]]
}

