% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/analyzeDuplicatedPeptides.R
\name{getUniqueEntries}
\alias{getUniqueEntries}
\title{get a list of of unique entries in rownames}
\usage{
getUniqueEntries(rownames)
}
\arguments{
\item{rownames}{list with duplicated entries which will be removed}
}
\description{
get a list of of unique entries in rownames
}
\examples{
rownames = c("a","b","a","c","d","d","e")
res = getUniqueEntries(rownames)
stopifnot(rownames[res] == c("b","c","e"))
}
\seealso{
\code{\link{analyzeDuplicated}} for contex
}

