% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotting.R
\name{imagelabels}
\alias{imagelabels}
\title{image plot with labels}
\usage{
imagelabels(x, labels = colnames(x), cex = 1, cex.axis = 0.5,
  main = NULL, col = heat.colors(12))
}
\arguments{
\item{x}{- matrix}

\item{labels}{- colnames(x)}

\item{cex}{- size of labels}

\item{cex.axis}{- size of axis lables}
}
\description{
image plot with labels
}
\examples{
x = matrix(rnorm(20*20),ncol=20)
imagelabels(x)
}

