% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\docType{package}
\name{imsbInfer}
\alias{imsbInfer}
\alias{imsbInfer-package}
\title{A package with some utility functions for the analysis of swath data}
\description{
A package with some utility functions for the analysis of swath data
}

