% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/loadTransitionsMSExperiment.R
\name{loadTransitonsMSExperiment}
\alias{loadTransitonsMSExperiment}
\title{load msexperiment with nrt transtions and peptides}
\usage{
loadTransitonsMSExperiment(obj, nrt = 3, peptop = 3)
}
\arguments{
\item{obj}{a table produced by feature alignment}

\item{nrt}{number of transitions}

\item{peptop}{nr of top peptides}
}
\description{
Selects top nrt transitions based on median transition intensity in all runs.
Selects top nr peptides based on median peptide intensity in all runs.
}
\examples{
library(imsbInfer)
data( feature_alignment_requant )

#SpecLib = ("C:/Users/witek/Google Drive/DataAnalysis/EBhardt/data/E1404301658-sample-SpecLib/feature_alignment_requant.tsv")
#SpecLib = ("/media/witold/data/googledrive/DataAnalysis/EBhardt/data/E1404301658-sample-SpecLib/feature_alignment_requant.tsv")
#obj  = fread(SpecLib)
#nrt = 3
#peptop = 3
#obj =feature_alignment_requant
x = loadTransitonsMSExperiment( feature_alignment_requant , nrt= 3, peptop=3)
x = loadTransitonsMSExperiment( feature_alignment_requant )
table(table(feature_alignment_requant$transition_group_id))
table(table(x$pepinfo$transition_group_id))
x2 = loadTransitonsMSExperiment(feature_alignment_requant, nrt= 20, peptop=20)
table(table(x2$pepinfo$ProteinName))
table(table(x2$pepinfo$PeptideSequence))
dim(x)
head(x$pepinfo)
mypairs(x$Intensity[,1:3])
#check that ordering is consistent
xx = split2table(rownames(x$pepinfo),split="-")
stopifnot(xx[,1] == x$pepinfo$transition_group_id)
stopifnot(xx[,2] == x$pepinfo$aggr_Fragment_Annotation)
}

