% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msExperiment.R
\docType{class}
\name{msTransitionsExperiment-class}
\alias{msTransitionExperiment}
\alias{msTransitionsExperiment-class}
\title{Holds Transition level data}
\description{
#@field transitiondata transtion data table
#@field precdata precursor data table
#@field columnsAll required columns for transitiondata
#@field columnsPrecursor required columns for precursor
}
\section{Methods}{

\describe{
\item{\code{getDecoy()}}{matrix with decoy information}

\item{\code{getFragmentIntensities()}}{matrix with transitions intensities}

\item{\code{getGlobalFDR()}}{compute FDR for dataset}

\item{\code{getPrecursorMZ()}}{matrix with precursor mz}

\item{\code{getPrecursorRT()}}{matrix with precursor retention times}

\item{\code{getPrecursorScore()}}{matrix with precursor scores}

\item{\code{plotSampleFDR(log = "")}}{plot FDR versus Score}

\item{\code{setData(data)}}{set the data}
}}
\examples{
rm(list=ls())
library(imsbInfer2)
library(readr)
data <- read_tsv("inst/extdata/example.tsv.gz",col_names = TRUE)
dim(data)
data <- prepareOpenSwathData(data)

huhu <- msTransitionExperiment()
huhu$setData(data)
intTrans <- huhu$getFragmentIntensities()
precRT <- huhu$getPrecursorRT()
dim(intTrans)
head(intTrans)
huhu$getGlobalFDR()
huhu$plotGlobalFDR()
decs<-huhu$getDecoy()
}

