% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/swathDataUtils.R
\name{mycolnames}
\alias{mycolnames}
\title{allows to set colnames for all the matrices in msexperiment}
\usage{
mycolnames(data, newNames)
}
\arguments{
\item{data}{data}

\item{newNames}{newNames}
}
\description{
allows to set colnames for all the matrices in msexperiment
}
\examples{
data(SDat)
SDat = mycolnames(SDat, c("bal1","bla2","bla3"))
}

