% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/swathDataUtils.R
\name{orderByKey.msexperiment}
\alias{orderByKey.msexperiment}
\title{orders rows by rowkey}
\usage{
\method{orderByKey}{msexperiment}(obj)
}
\arguments{
\item{experiment}{obj}
}
\description{
orders rows by rowkey
}
\examples{
data(SDat)
head(SDat$rt)
SDat=orderByKey(SDat)
rownames(SDat$Intensity)
}

