% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/swathDataUtils.R
\name{orderByRT.msexperiment}
\alias{orderByRT.msexperiment}
\title{orders rows by retention time of peptide}
\usage{
\method{orderByRT}{msexperiment}(obj)
}
\arguments{
\item{obj}{experiment}
}
\description{
orders rows by retention time of peptide
}
\examples{
data(SDat)
head(SDat$rt)
SDat=orderByRT(SDat)
head(SDat$rt)
SDat$RT[1:10]
}

