% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotting.R
\name{pairsDifference}
\alias{pairsDifference}
\title{plot x-y against retention time}
\usage{
pairsDifference(dataf, RT, main = "", ylim = NULL, maxPanel = 8)
}
\arguments{
\item{dataf}{dataframe}

\item{RT}{retention time vector}

\item{main}{title}

\item{ylim}{axis extend}

\item{maxPanel}{maximum number of datasets to display. dim(dataf)[2] > maxPanel, maxPanel columns are sampled}
}
\description{
plot x-y against retention time
}
\examples{
data(SDat)
rtord = orderByRT(SDat)
pairsDifference( asinh(rtord$Intensity), rtord$RT , ylim=NULL)
}
\seealso{
\code{\link{pairs}} and \code{\link{pairsRatio}}
}

