% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotting.R
\name{pairsQQ}
\alias{pairsQQ}
\title{pairsplot of QQ plots}
\usage{
pairsQQ(dataframesel, main = "")
}
\description{
pairsplot of QQ plots
}
\examples{
data(SDat)
pairsQQ( SDat$Intensity )
}
\seealso{
\code{\link{qqplot}} and  \code{\link{pairs}}
}

