% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotting.R
\name{pairsRatio}
\alias{pairsRatio}
\title{plot log2(x/y) against retention time}
\usage{
pairsRatio(dataframesel, RT, main = "", ylim = c(-2, 2))
}
\description{
plot log2(x/y) against retention time
}
\examples{
data(SDat)
rtord = orderByRT(SDat)
pairsRatio( rtord$Intensity, rtord$RT )
}
\seealso{
\code{\link{pairs}} and \code{\link{pairsDifference}}
}

