% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openSwath.R
\name{prepareOpenSwathData}
\alias{prepareOpenSwathData}
\title{Prepare data from OpenSwathOutput}
\usage{
prepareOpenSwathData(far)
}
\description{
Prepare data from OpenSwathOutput
}
\examples{
library(readr)
library(imsbInfer2)
file = "d:/GoogleDrive/tissuecomparison/OpenSWATH/BAT_19strains/data/E1603291025_feature_alignment_requant.tsv.gz"
file = "C:/Users/wewol/Google Drive/tissuecomparison/OpenSWATH/BAT_19strains/data/E1603291025_feature_alignment_requant.tsv.gz"
data <- read_tsv(file,col_names = TRUE)

far <- .prepareDF(data)
data3 <- prepareOpenSwathData(data)

}

