% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/analyzeDuplicatedPeptides.R
\name{randomPairs}
\alias{randomPairs}
\title{sample pairs from pepidx}
\usage{
randomPairs(n, pepidx)
}
\arguments{
\item{n}{number of pairs}

\item{pepidx}{which peptides to sample}
}
\description{
sample pairs from pepidx
}
\examples{
data(SDat)
res = randomPairs(100, 1:dim(SDat)[1])
length(res)
res[[1]]
res2 = analyzeDuplicated( SDat , res)
dim(res2)
head(res2)
}
\seealso{
\code{\link{analyzeDuplicated}}
}

