% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swathDataUtils.R
\name{read2msExperiment.data.frame}
\alias{read2msExperiment.data.frame}
\title{convert data.frame 2 msexperiment}
\usage{
\method{read2msExperiment}{data.frame}(obj, ...)
}
\arguments{
\item{obj}{a data.frame in long format}

\item{...}{unused}
}
\value{
msexperiment
}
\description{
convert data.frame 2 msexperiment
}
\examples{
data(feature_alignment_requant)
obj=feature_alignment_requant
SDat = read2msExperiment(feature_alignment_requant)
stopifnot(dim(SDat)==c(964,3))
stopifnot(rownames(SDat$pepinfo)==rownames(SDat$Intensity))
\dontrun{save(SDat,file="data/SDat.rda")}
}

