% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/swathDataUtils.R
\name{removeDecoys.msexperiment}
\alias{removeDecoys.msexperiment}
\title{remove decoys and decoy column from msexperiment}
\usage{
\method{removeDecoys}{msexperiment}(obj, ...)
}
\arguments{
\item{obj}{msexperiment}

\item{...}{unused}
}
\description{
remove decoys and decoy column from msexperiment
}
\examples{
data(feature_alignment_requant)
SDat = read2msExperiment(feature_alignment_requant)
dim(SDat)
SDat = removeDecoys(SDat)
lapply(SDat,dim)
}

