% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swathDataUtils.R
\name{removeSample}
\alias{removeSample}
\title{remove sample from dataset}
\usage{
removeSample(data, sample)
}
\arguments{
\item{data}{msexperiment}

\item{sample}{which sample to remove}
}
\description{
remove sample from dataset
}
\examples{
data(SDat)
new = removeSample(SDat, sample="chludwig_L110830_20_SW")
colnames(new$Intensity)
new = removeSample(SDat, sample=NULL)
colnames(new$Intensity)
}

