% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swathDataUtils.R
\name{removeSamples}
\alias{removeSamples}
\title{remove a specific samples from the dataset}
\usage{
removeSamples(data, samples)
}
\arguments{
\item{data}{msexperiment}

\item{samples}{which samples to remove}
}
\description{
remove a specific samples from the dataset
}
\examples{

data(SDat)
colnames(SDat$Intensity)
new = removeSamples(SDat, sample=c("chludwig_L110830_20_SW","chludwig_L110830_22_SW"))
colnames(new$Intensity)
new = removeSamples(SDat, NULL)
colnames(new$Intensity)

}

