% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/swathDataUtils.R
\name{reordercolumns}
\alias{reordercolumns}
\title{reorders all the matrices by the columnnames}
\usage{
reordercolumns(data, ord = NULL)
}
\arguments{
\item{data}{msexperiment to reorder}

\item{ord}{new ordering - if null then use ordering by column names}
}
\description{
reorders all the matrices by the columnnames
}
\examples{
data(SDat)
SDat = mycolnames(SDat, c("C","A","B"))
dimnames(SDat)
SDat = reordercolumns(SDat)
dimnames(SDat)
}

