% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utilities.R
\name{rmNArows}
\alias{rmNArows}
\title{Removes rows with NA's from matrix}
\usage{
rmNArows(plev, thresh = 0)
}
\value{
matrix
}
\description{
Removes rows with NA's from matrix
}
\examples{
x = matrix(rnorm(10*10),ncol=10)
dim(x)
x[3,3] = NA
x = rmNArows(x)
dim(x)
}

