% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\docType{data}
\name{SDat}
\alias{SDat}
\title{SDat - example of an msexperiment class}
\format{see example below}
\usage{
feature_alignment_requant
}
\description{
this is an example output of calling the function read2msExperiment
}
\examples{
data(feature_alignment_requant)
SDat = read2msExperiment(feature_alignment_requant)
names(SDat)
}
\author{
Witold Wolski
}
\seealso{
\code{\link{feature_alignment_requant}} and \code{\link{read2msExperiment.data.frame}}
}

