% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transitionAnalysis.R
\name{selectTopFragmentsPerPeptide}
\alias{selectTopFragmentsPerPeptide}
\title{selecting top transtions WARNING - slow running function}
\usage{
selectTopFragmentsPerPeptide(data, nrt = 2)
}
\arguments{
\item{data}{- data.frame or table with at least 2 columns:  transition_group_id, aggr_Fragment_Annotation (unique keys)}

\item{nrt}{- top transitions to select}
}
\description{
selecting top transtions WARNING - slow running function
}
\examples{
data(feature_alignment_requant)
table(table(feature_alignment_requant$transition_group_id))
tmp = transitions2wide(feature_alignment_requant)
# how many peptides with how many transitions
x1 = table(table(tmp$transition_group_id))
xx = selectTopFragmentsPerPeptide(tmp,nrt=7)
x2 = table(table(xx$transition_group_id))
# because with nrt 7 you do not remove any transitions
stopifnot(x1 == x2)
xx = selectTopFragmentsPerPeptide(tmp,nrt=2)
# all peptides must have 2 transitions
stopifnot(names(table(table(xx$transition_group_id))) == "2")
}

