% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transitionAnalysis.R
\name{selectTopPeptidesPerProtein}
\alias{selectTopPeptidesPerProtein}
\title{this function selects the top x peptides / protein}
\usage{
selectTopPeptidesPerProtein(msexp, peptop = 3)
}
\arguments{
\item{msexp}{data.frame with}

\item{peptop}{how many top peptides}
}
\description{
this function selects the top x peptides / protein
}
\examples{
data(SDat)
rownames(SDat$pepinfo)
table(table(SDat$pepinfo$ProteinName))
x = selectTopPeptidesPerProtein(SDat,peptop=3)
table(table(x$pepinfo$ProteinName))
stopifnot(rownames(x$pepinfo)[1:10]==rownames(x$Intensity)[1:10])
stopifnot( length(unique(SDat$pepinfo$ProteinName)) == length(unique(x$pepinfo$ProteinName)) )
}

