% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swathDataUtils.R
\name{subset.msexperiment}
\alias{subset}
\alias{subset.msexperiment}
\title{subset data given idx vector}
\usage{
\method{subset}{msexperiment}(x, idx, ...)
}
\arguments{
\item{x}{msexperiment}

\item{idx}{row indices to keep or of TRUE, FALSE vector}

\item{...}{unused}
}
\description{
subset data given idx vector
}
\examples{
data(SDat)
# keep only peakgroups with mass less than 800
mass = apply(SDat$mz,1,median)
dim(SDat)
SDatr = subset(SDat, mass < 800)
dim(SDatr)
}

