% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/transitionAnalysis.R
\name{transitions2wide}
\alias{transitions2wide}
\title{extract transition intensities from input file. WARNING - slow running function}
\usage{
transitions2wide(far)
}
\arguments{
\item{far}{convert to wide format}
}
\description{
the input table required fields are:
transition_group_id,
aggr_Peak_Area,
aggr_Fragment_Annotion,
align_origfilename,
}
\examples{
data(feature_alignment_requant)
df = prepareDF(feature_alignment_requant)
tmp = transitions2wide(df)
names(df)
dim(feature_alignment_requant)
dim(tmp)[1]/dim(feature_alignment_requant)[1]*3
}

