% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotting.R
\name{volcanoplot}
\alias{volcanoplot}
\title{volcano plot}
\usage{
volcanoplot(foldchange, pvals, pthresh = 0.05, ratiothresh = 2,
  xlab = "log2(T/N)", ylab = "-log10(P)", labels = NULL, cex = 0.6,
  xlim = NULL)
}
\arguments{
\item{foldchange}{- fold change values}

\item{pvals}{pvalues}

\item{pthresh}{pvalue threshold}

\item{ratiothresh}{threshold of foldchange}

\item{cex}{size of labels}

\item{labals}{- optional labels}
}
\description{
volcano plot
}
\examples{
foldchange <- rnorm(1000)
pval <-rexp(1000)
volcanoplot(foldchange, pval)
}

