% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregateMSexperiment.R
\name{aggregatePeptides}
\alias{aggregatePeptides}
\title{aggregate peptides}
\usage{
aggregatePeptides(msexp, FUN = sum)
}
\arguments{
\item{msexp}{msexperiment object}

\item{FUN}{aggregation function}
}
\description{
Given msexperiment with transitons aggregate the transitions to produce precursor intensities
}
\examples{
data(feature_alignment_requant)
colnames(feature_alignment_requant)
dim(feature_alignment_requant)/3
msexp = loadTransitonsMSExperiment( feature_alignment_requant , nrt = 3 , peptop = 3 )
dim(msexp)
sum(rownames(msexp$Intensity) == rownames(msexp$pepinfo))
y = aggregatePeptides(msexp)
sum(rownames(y$Intensity) == rownames(y$pepinfo))
dim(y)
stopifnot(table(table(y$pepinfo$transition_group_id)) == 247)
}
\seealso{
aggregateProtein
}

