% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msExperiment.R
\docType{class}
\name{msTransitionExperiment-class}
\alias{msTransitionExperiment}
\alias{msTransitionExperiment-class}
\title{Holds Transition level data}
\description{
Holds Transition level data
}
\section{Fields}{

\describe{
\item{\code{name}}{name of experiment}

\item{\code{path}}{where to store data}

\item{\code{peptide}}{transtion data table}

\item{\code{precursor}}{precursor data table}

\item{\code{transition}}{required columns for transitiondata}
}}
\section{Methods}{

\describe{
\item{\code{getFragmentIntensities()}}{matrix with transitions intensities}

\item{\code{getGlobalFDR()}}{compute FDR for dataset}

\item{\code{getPrecursorIntensity(FUN = sum)}}{!!!don't use. work in progress!!!}

\item{\code{getPrecursorMZ()}}{matrix with precursor mz}

\item{\code{getPrecursorRT()}}{matrix with precursor retention times}

\item{\code{getPrecursorScore()}}{matrix with precursor scores}

\item{\code{plotSampleFDR(log = "")}}{plot FDR versus Score}

\item{\code{setData(data, IsotopeLabelType = "L")}}{set the data}
}}
\examples{
rm(list=ls())
library(imsbInfer2)
library(readr)
library(dplyr)
data <- read_tsv("inst/extdata/example.tsv.gz",col_names = TRUE)
data <- prepareOpenSwathData(data)

huhu <- msTransitionExperiment()
huhu$setData(data)
data <- huhu$getData() 
x<-huhu$peptide
head(x)
x <- huhu$
dumm <- msTransitionExperiment(path=".", name="mydb.sql")
dumm$name
dumm$setData(data)
huhu <- dumm
intTrans <- huhu$getFragmentIntensities()
dim(intTrans)
huhu$noDecoy()
colnames(intTrans)
intTrans <- huhu$getFragmentIntensities()
dim(intTrans)
pairs(intTrans[,5:ncol(intTrans)],log="xy",pch=".")
precRT <- huhu$getPrecursorRT()
precMZ <- huhu$getPrecursorMZ()
precScore <- huhu$getPrecursorScore()
pairs(precScore[,3:ncol(precScore)],log="xy",pch=".")
colnames(huhu$precursor)
colnames(huhu$peptide)

head(huhu$peptide[huhu$peptide$Decoy==1,])
huhu$withDecoy()
head(huhu$peptide[huhu$peptide$Decoy==1,])
huhu$noDecoy()
length(unique(huhu$peptide$StrippedSequence))
length(unique(huhu$peptide$ProteinName))
xx <-merge(huhu$peptide[,c("StrippedSequence","Decoy")], huhu$precursor )

test <- (huhu$getPrecursorIntensity())
huhu$getGlobalFDR()
decs<-huhu$getDecoy()

}

