% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msExperiment.R
\docType{class}
\name{msTransitionExperiment-class}
\alias{msTransitionExperiment}
\alias{msTransitionExperiment-class}
\title{Holds Transition level data}
\description{
Holds Transition level data
}
\section{Fields}{

\describe{
\item{\code{name}}{name of experiment}

\item{\code{path}}{where to store data}

\item{\code{peptide}}{transtion data table}

\item{\code{precursor}}{precursor data table}

\item{\code{transition}}{required columns for transitiondata}

\item{\code{labelType}}{Stores only one label type}
}}
\section{Methods}{

\describe{
\item{\code{getData()}}{get all the data in long format}

\item{\code{getFileName()}}{get filenames}

\item{\code{getFragmentIntensities()}}{matrix with transitions intensities}

\item{\code{getGlobalFDR()}}{compute FDR for dataset}

\item{\code{getPrecursorIntensity()}}{DO NOT USE!!! Selects Fragment intensities given by external software, }

\item{\code{getPrecursorMZ()}}{matrix with precursor mz}

\item{\code{getPrecursorRT()}}{matrix with precursor retention times}

\item{\code{getPrecursorScore()}}{matrix with precursor scores}

\item{\code{noDecoy()}}{work in decoy free mode}

\item{\code{plotFalsePostives(log = "")}}{plot FP versus Score}

\item{\code{setData(data, labelType = "L")}}{set the data (drops all data if already existing)}

\item{\code{withDecoy()}}{work with decoys}
}}
\examples{
rm(list=ls())
library(imsbInfer2)
library(readr)
library(dplyr)
data <- read_tsv(file.path(path.package("imsbInfer2"),"extdata/example.tsv.gz"),col_names = TRUE)
data <- prepareOpenSwathData(data)

huhu <- msTransitionExperiment(path=".", name="mydb2.sql")
huhu$setData(data)
tmp <- huhu$getData()
stopifnot(dim(tmp) == dim(data))
huhu$getFileName()

intTrans <- huhu$getFragmentIntensities()
dim(intTrans)
head(intTrans)

precIntExt <- huhu$getPrecursorIntensity()
head(precIntExt)
dim(precIntExt)
huhu$noDecoy()

precRT <- huhu$getPrecursorRT()
precMZ <- huhu$getPrecursorMZ()
precScore <- huhu$getPrecursorScore()
head(precScore)
pairs(precScore[,6:ncol(precScore)],log="xy",pch=".")
dim(huhu$precursor)
dim(huhu$precursorId)
table(huhu$peptide$Freq)

head(huhu$peptide[huhu$peptide$Decoy==1,])
huhu$withDecoy()
head(huhu$peptide[huhu$peptide$Decoy==1,])
dim(huhu$precursor)
dim(huhu$precursorId)
huhu$noDecoy()
length(unique(huhu$peptide$StrippedSequence))
length(unique(huhu$peptide$ProteinName))
xx <-merge(huhu$peptide[,c("StrippedSequence","Decoy")], huhu$precursor )
head(xx)
huhu$withDecoy()
huhu$getGlobalFDR()
huhu$plotFalsePostives()
}

