% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msExperiment.R
\docType{class}
\name{msTransitions-class}
\alias{msTransitions}
\alias{msTransitions-class}
\title{Holds Transition level data}
\description{
Holds Transition level data
}
\section{Fields}{

\describe{
\item{\code{transitiondata}}{transtion data table}

\item{\code{precdata}}{precursor data table}

\item{\code{columnsAll}}{required columns for transitiondata}

\item{\code{columnsPrecursor}}{required columns for precursor}
}}
\section{Methods}{

\describe{
\item{\code{getMZ()}}{matrix with precursor mz}

\item{\code{getRT()}}{matrix with precursor retention times}

\item{\code{getScore()}}{matrix with precursor scores}

\item{\code{getTransIntensity()}}{matrix with transitions intensities}

\item{\code{setData(data)}}{set the data}
}}
\examples{
huhu <- msTransitions()
huhu$columnsAll
huhu$columnsPrecursor
huhu$transitiondata
file = file.path(path.package("imsbInfer2"),"extdata/E1603291025_feature_alignment_requant.tsv.gz")
data <- read_tsv(file,col_names = TRUE)
data <- prepareOpenSwathData(data)
head(data)
huhu$setData(data)
intensity <- huhu$getTransIntensity()
dim(intensity)
rt <-huhu$getRT()
}

