% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{pairsRatio}
\alias{pairsRatio}
\title{plot log2(x/y) against retention time}
\usage{
pairsRatio(data, RT, main = "", ylim = c(-2, 2))
}
\arguments{
\item{data}{data.frame with intensities}

\item{RT}{rt information (same langth as nrow(data))}

\item{main}{main title}

\item{ylim}{y image extend}
}
\description{
plot log2(x/y) against retention time
}
\examples{
data(SDat)
rtord = orderByRT(SDat)
pairsRatio( rtord$Intensity, rtord$RT )
}
\seealso{
\code{\link{pairs}} and \code{\link{pairsDifference}}
}

