% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/analyzeDuplicatedPeptides.R
\name{plotPairCors}
\alias{plotPairCors}
\title{plot the output of compPeptideCorrelations}
\usage{
plotPairCors(res, main = "", xlim = c(-1, 1))
}
\arguments{
\item{res}{result of compPeptideCorrelations}

\item{main}{title}

\item{xlim}{image range}
}
\description{
plot the output of compPeptideCorrelations
}

