% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transitionAnalysis.R
\name{prepareDF}
\alias{prepareDF}
\title{checks input file for required columns - removes nonrequired columns from data.frame}
\usage{
prepareDF(df)
}
\arguments{
\item{df}{ouptut fo OpenSwath feature aligner}
}
\description{
checks input file for required columns - removes nonrequired columns from data.frame
}
\examples{
library(imsbInfer)
data(feature_alignment_requant)
colnames(feature_alignment_requant)
df = feature_alignment_requant
df = prepareDF(feature_alignment_requant)
colnames(df)
length(unique(df$ProteinName))
}

