% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapOpenSwath.R
\name{prepareOpenSwathData}
\alias{prepareOpenSwathData}
\title{Prepare data from OpenSwathOutput}
\usage{
prepareOpenSwathData(far)
}
\description{
Prepare data from OpenSwathOutput
}
\examples{
rm(list=ls())
library(readr)
library(imsbInfer2)
data2 <- read_tsv(file.path(path.package("imsbInfer2"),"extdata/example.tsv.gz"),col_names = TRUE)
head(data2)

#far <- .prepareDF(data2)
#far \%>\% glimpse
#head(far)
#tmp[1:3]
data3 <- prepareOpenSwathData(data2)
colnames(data3)
dim(data3)

}

